removed external references to jquery or other things
for the index page, we can't link to libraries hosted elsewhere, as we want this page to be very reliable, we also don't want google to know what we are doing, and we also want to keep the page as lean as possible and close the the w3c standards.
for the admin page, it doesn't matter, as it's not the main page.